+++ src/pages/user_albums/user_albums.tsx
@@ -1,7 +1,5 @@
-import styles from './user_albums.module.scss';
 import { useNavigate, useParams } from 'react-router-dom';
-import { Album } from '../../entities/album/ui/Album.tsx';
-import { useGetAlbumsByUserIdQuery } from '../../entities/album/api/albumsApi.ts';
+import { AlbumsList } from '../../widgets/AlbumsList/AlbumsList.tsx';
 
 export const UserPageAlbums = () => {
   const { id: userId } = useParams();
@@ -9,19 +7,5 @@ export const UserPageAlbums = () => {
 
   if (!userId) navigate('/posts');
 
-  const { data: userAlbums = [], isLoading } = useGetAlbumsByUserIdQuery(
-    userId!
-  );
-
-  return isLoading ? (
-    <h3>Загрузка...</h3>
-  ) : (
-    <div className={styles.albumsListWrapper}>
-      <ul className={styles.albumsList}>
-        {userAlbums.map((album) => (
-          <Album album={album} key={album.id} />
-        ))}
-      </ul>
-    </div>
-  );
+  return <AlbumsList userId={userId} />;
 };
