+++ src/pages/user_todos/user_todos.tsx
@@ -1,7 +1,5 @@
-import styles from './user_todos.module.scss';
 import { useNavigate, useParams } from 'react-router-dom';
-import { Todo } from '../../entities/todo/ui/Todo.tsx';
-import { useGetTodosByUserIdQuery } from '../../entities/todo/api/todosApi.ts.ts';
+import { TodosList } from '../../widgets/TodosList/TodosList.tsx';
 
 export const UserPageTodos = () => {
   const { id: userId } = useParams();
@@ -9,19 +7,5 @@ export const UserPageTodos = () => {
 
   if (!userId) navigate('/posts');
 
-  const { data: userTodos = [], isLoading } = useGetTodosByUserIdQuery(userId!);
-
-  return isLoading ? (
-    <h3>Загрузка...</h3>
-  ) : (
-    <div className={styles.todosListWrapper}>
-      <ul className={styles.todosList}>
-        {userTodos.map((todo) => (
-          <li className={styles.todoItem} key={todo.id}>
-            <Todo todo={todo} />
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
+  return <TodosList userId={userId} />;
 };
