+++ src/widgets/PhotosList/PhotosList.module.scss
